+++ test/viking/model/class_properties/reflect_on_associations.js
@@ -5,7 +5,7 @@
         Ship = Viking.Model.extend({ hasMany: ['ships'], belongsTo: [['carrier', {model: 'Ship'}]] });
         ShipCollection = Viking.Collection.extend({ model: Ship });
         
-        deepEqual(['ships', 'carrier'], _.map(Ship.reflectOnAssociations(), function(a) { return a.name; }));
+        deepEqual(['carrier', 'ships'], _.map(Ship.reflectOnAssociations(), function(a) { return a.name; }));
     
         delete Ship;
         delete ShipCollection;